feat: Add ability to convert to double
Add the ability to convert from string to double.

Worth noting is the fact that the conversion is locality sensitive.
This means that for a decimalseparator a ',' (comma), since I live in
Sweden, was needed to make an accepted conversion. And not a '.'.
https://docs.microsoft.com/en-us/dotnet/api/system.double.tryparse?view=net-6.0

This implementation lacks the numberstyles feature
https://docs.microsoft.com/en-us/dotnet/api/system.globalization.numberstyles?view=net-6.0

Fixes #8