refactor: Move common Verify code to a baseclass
All verify classes have the same codestructure. So I moved it to a
baseclass to make it easier to implement and reduce risk of bugs.

The verify-classes now only have to call the BaseVerify.VerifyValue()

Sadly you cannot inherit BaseVerify from the Verify-classes, so the
BaseVerify.VerifyValue() is standalone and static.

Fixes #11