feat: Add StringModifier, w/ Padding and Truncate
Add StringModifier class with the methods Padding and Truncate.
Padding uses the method Truncate, hence commiting both at the same time.

With Padding you can add extra characters before and/or after
your string in order to make it look better aligned.
"Hello" => "Hello..." (Padded with '.' to give it a width of 8)

Truncate cuts of the remaning characters in the string if it excedes
the max length.